+++ wiki/wiki_md/about/changelog.md
@@ -12,6 +12,7 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 ### Changes
 
 - Added a config option to configure the maximum number of animals allowed within the Nature Core's Animal Spawning radius.
+- Added the ability to configure the Spectre Energy Injector capacity, Spectre Coil/Charger transfer rates, and whether the Genesis Spectre Coil generates energy or transfers it.
 
 ### Fixes